+++ CHANGELOG.md
@@ -8,7 +8,7 @@
 - Tetris preview upcoming tetromino shape
 - Added parent_proxy to CustomWidget and CustomShell classes
 - Update CustomShell#center and ShellProxy#center to center_within_display to avoid clash with `row_layout` center property
-- Fixed issue with shell/dialog/custom-shell not maintaining parent when not passed and is derived automatically from last open shell
+- Fixed issue with shell/dialog/custom-shell not maintaining parent when not passed
 - Fix Tetris sideways edge detection
 
 ### 4.18.2.3